requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.1.0
google-generativeai==0.3.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
PyJWT==2.8.0
bcrypt==4.0.1
langdetect==1.0.9
reportlab==4.0.4
markdown==3.5.1
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0 